+++ src/dotnet/APIView/ClientSPA/src/app/_components/conversation-page/conversation-page.component.scss
@@ -1,0 +1,6 @@
+:host ::ng-deep {
+    .conversation-panel{
+        height: calc(100vh - 130px);
+        background-color: var(--base-fg-color);
+    }
+}
