+++ src/dotnet/APIView/ClientSPA/src/app/_components/review-page/review-page.component.spec.ts
@@ -16,7 +16,7 @@ import { ReviewNavComponent } from '../review-nav/review-nav.component';
 import { ReviewPageOptionsComponent } from '../review-page-options/review-page-options.component';
 import { PageOptionsSectionComponent } from '../shared/page-options-section/page-options-section.component';
 import { SharedAppModule } from 'src/app/_modules/shared/shared-app.module';
-import { ReviewPageModule } from 'src/app/_modules/review-page/review-page.module';
+import { ReviewPageModule } from 'src/app/_modules/review-page.module';
 import { MessageService } from 'primeng/api';
 
 describe('ReviewPageComponent', () => {
