+++ src/dotnet/APIView/ClientSPA/src/app/_modules/conversation-page.module.ts
@@ -1,0 +1,22 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { ConversationPageComponent } from 'src/app/_components/conversation-page/conversation-page.component';
+import { ReviewPageLayoutModule } from './shared/review-page-layout.module';
+import { SharedAppModule } from './shared/shared-app.module';
+import { CommonModule } from '@angular/common';
+
+const routes: Routes = [
+  { path: '', component: ConversationPageComponent }
+];
+
+@NgModule({
+  declarations: [
+    ConversationPageComponent
+  ],
+  imports: [
+    CommonModule,
+    ReviewPageLayoutModule,
+    RouterModule.forChild(routes),
+  ]
+})
+export class ConversationPageModule { }
