+++ src/dotnet/APIView/ClientSPA/src/app/_modules/review-page.module.ts
@@ -1,0 +1,40 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { RouterModule, Routes } from '@angular/router';
+import { ReviewPageComponent } from 'src/app/_components/review-page/review-page.component';
+import { ReviewNavComponent } from 'src/app/_components/review-nav/review-nav.component';
+import { CodePanelComponent } from 'src/app/_components/code-panel/code-panel.component';
+import { DialogModule } from 'primeng/dialog';
+import { TreeSelectModule } from 'primeng/treeselect';
+import { ButtonModule } from 'primeng/button';
+import { UiScrollModule  } from 'ngx-ui-scroll' ;
+import { PageOptionsSectionComponent } from 'src/app/_components/shared/page-options-section/page-options-section.component';
+import { ReviewPageOptionsComponent } from 'src/app/_components/review-page-options/review-page-options.component';
+import { InputSwitchModule } from 'primeng/inputswitch';
+import { SharedAppModule } from './shared/shared-app.module';
+import { ReviewPageLayoutModule } from './shared/review-page-layout.module';
+
+const routes: Routes = [
+  { path: '', component: ReviewPageComponent }
+];
+
+@NgModule({
+  declarations: [
+    ReviewPageComponent,
+    ReviewNavComponent,
+    CodePanelComponent,
+    PageOptionsSectionComponent,
+    ReviewPageOptionsComponent,
+  ],
+  imports: [
+    CommonModule,
+    DialogModule,
+    TreeSelectModule,
+    ButtonModule,
+    InputSwitchModule,
+    UiScrollModule,
+    ReviewPageLayoutModule,
+    RouterModule.forChild(routes),
+  ]
+})
+export class ReviewPageModule { }
