+++ src/dotnet/APIView/ClientSPA/src/ng-prime-overrides.scss
@@ -376,7 +376,7 @@ p-contextmenusub {
 }
 
 .p-multiselect-panel .p-multiselect-items .p-multiselect-item:not(.p-highlight):not(.p-disabled).p-focus {
-    background: var(--base-fg-color);;
+    background: var(--base-fg-color);
 }
 
 .p-datatable .p-sortable-column.p-highlight, .p-datatable .p-sortable-column.p-highlight:hover {
